Refactoring to manage all values with uint64.

+++ config/config.go
@@ -55,16 +55,16 @@ type All struct {
 				Duration int  `yaml:"duration"`
 			}
 			Connections struct {
-				Enable   bool `yaml:"enable"`
-				Warning  int  `yaml:"warning"`
-				Critical int  `yaml:"critical"`
-				Duration int  `yaml:"duration"`
+				Enable   bool   `yaml:"enable"`
+				Warning  uint64 `yaml:"warning"`
+				Critical uint64 `yaml:"critical"`
+				Duration int    `yaml:"duration"`
 			}
 			Replication struct {
-				Enable   bool `yaml:"enable"`
-				Warning  int  `yaml:"warning"`
-				Critical int  `yaml:"critical"`
-				Duration int  `yaml:"duration"`
+				Enable   bool   `yaml:"enable"`
+				Warning  uint64 `yaml:"warning"`
+				Critical uint64 `yaml:"critical"`
+				Duration int    `yaml:"duration"`
 			}
 		}
 	}
@@ -77,10 +77,10 @@ type All struct {
 		}
 		Alerts struct {
 			Errors struct {
-				Enable   bool `yaml:"enable"`
-				Warning  int  `yaml:"warning"`
-				Critical int  `yaml:"critical"`
-				Duration int  `yaml:"duration"`
+				Enable   bool   `yaml:"enable"`
+				Warning  uint64 `yaml:"warning"`
+				Critical uint64 `yaml:"critical"`
+				Duration int    `yaml:"duration"`
 			}
 			Status struct {
 				Enable   bool `yaml:"enable"`
@@ -110,22 +110,22 @@ type All struct {
 		}
 		Alerts struct {
 			CPU struct {
-				Enable   bool `yaml:"enable"`
-				Warning  int  `yaml:"warning"`
-				Critical int  `yaml:"critical"`
-				Duration int  `yaml:"duration"`
+				Enable   bool   `yaml:"enable"`
+				Warning  uint64 `yaml:"warning"`
+				Critical uint64 `yaml:"critical"`
+				Duration int    `yaml:"duration"`
 			}
 			Disk struct {
-				Enable   bool `yaml:"enable"`
-				Warning  int  `yaml:"warning"`
-				Critical int  `yaml:"critical"`
-				Duration int  `yaml:"duration"`
+				Enable   bool   `yaml:"enable"`
+				Warning  uint64 `yaml:"warning"`
+				Critical uint64 `yaml:"critical"`
+				Duration int    `yaml:"duration"`
 			}
 			MEM struct {
-				Enable   bool `yaml:"enable"`
-				Warning  int  `yaml:"warning"`
-				Critical int  `yaml:"critical"`
-				Duration int  `yaml:"duration"`
+				Enable   bool   `yaml:"enable"`
+				Warning  uint64 `yaml:"warning"`
+				Critical uint64 `yaml:"critical"`
+				Duration int    `yaml:"duration"`
 			}
 		}
 	}
